adding not yet implemented to things

+++ mech/mech.py
@@ -381,7 +381,6 @@ class Mech(MechCommand):
                         print(colored.yellow("VM ({}) was already started on an "
                                              "unknown IP address".format(instance)))
 
-                # TODO: add these for virtualbox
                 if inst.provider == 'vmware':
                     # if not already using preshared key, switch to it
                     if not inst.use_psk and inst.auth:
@@ -393,6 +392,9 @@ class Mech(MechCommand):
 
                     if not disable_provisioning:
                         utils.provision(inst, show=False)
+                else:
+                    print(colored.red("Not yet implemented on this platform:"))
+                    print(colored.red("(use-psk, remove-vagrant, provisioning)"))
 
     # allows "mech start" to alias to "mech up"
     start = up
@@ -407,7 +409,14 @@ class Mech(MechCommand):
             -h, --help                       Print this help
         """
         vmrun = VMrun()
-        print(vmrun.list())
+        if vmrun.installed():
+            print("===VMware VMs===")
+            print(vmrun.list())
+
+        vbm = VBoxManage()
+        if vbm.installed():
+            print("===VirtualBox VMs===")
+            print(vbm.list())
 
     def ps(self, arguments):  # pylint: disable=invalid-name,no-self-use
         """
@@ -423,9 +432,13 @@ class Mech(MechCommand):
         inst = MechInstance(instance_name)
 
         if inst.created:
-            # Note: user/password is needed for ps
-            vmrun = VMrun(inst.vmx, inst.user, inst.password)
-            print(vmrun.list_processes_in_guest())
+            if inst.provider == 'vmware':
+                # Note: user/password is needed for ps
+                vmrun = VMrun(inst.vmx, inst.user, inst.password)
+                print(vmrun.list_processes_in_guest())
+            else:
+                print(colored.red("Not yet implemented on this platform."))
+
         else:
             print("VM {} not created.".format(instance_name))
 
@@ -454,27 +467,32 @@ class Mech(MechCommand):
             inst = MechInstance(instance)
 
             if inst.created:
-                vmrun = VMrun(inst.vmx)
 
-                ip_address = inst.get_ip()
-                state = vmrun.check_tools_state(quiet=True)
+                if inst.provider == 'vmware':
+                    vmrun = VMrun(inst.vmx)
 
-                print("Current machine state:" + os.linesep)
-                if ip_address is None:
-                    ip_address = "poweroff"
-                elif not ip_address:
-                    ip_address = "unknown"
-                print("%s\t%s\t%s\t(VMware Tools %s)" % (inst.name, inst.box, ip_address, state))
+                    ip_address = inst.get_ip()
+                    state = vmrun.check_tools_state(quiet=True)
 
-                if ip_address == "poweroff":
-                    print(os.linesep + "The VM is powered off. To restart the VM, "
-                          "simply run `mech up {}`".format(instance))
-                elif ip_address == "unknown":
-                    print(os.linesep + "The VM is on. but it has no IP to connect to,"
-                          "VMware Tools must be installed")
-                elif state in ("installed", "running"):
-                    print(os.linesep + "The VM is ready. Connect to it "
-                          "using `mech ssh {}`".format(instance))
+                    print("Current machine state:" + os.linesep)
+                    if ip_address is None:
+                        ip_address = "poweroff"
+                    elif not ip_address:
+                        ip_address = "unknown"
+                    print("%s\t%s\t%s\t(VMware Tools %s)" % (inst.name, inst.box,
+                                                             ip_address, state))
+
+                    if ip_address == "poweroff":
+                        print(os.linesep + "The VM is powered off. To restart the VM, "
+                              "simply run `mech up {}`".format(instance))
+                    elif ip_address == "unknown":
+                        print(os.linesep + "The VM is on. but it has no IP to connect to,"
+                              "VMware Tools must be installed")
+                    elif state in ("installed", "running"):
+                        print(os.linesep + "The VM is ready. Connect to it "
+                              "using `mech ssh {}`".format(instance))
+                else:
+                    print(colored.red("Not yet implemented on this platform."))
             else:
                 print("The VM ({}) has not been created.".format(instance))
 
@@ -596,11 +614,15 @@ class Mech(MechCommand):
             inst = MechInstance(instance)
 
             if inst.created:
-                vmrun = VMrun(inst.vmx)
-                if vmrun.pause() is None:
-                    print(colored.red("Not paused", vmrun))
+                if inst.provider == 'vmware':
+                    vmrun = VMrun(inst.vmx)
+                    pause_results = vmrun.pause()
+                    if pause_results is None:
+                        print(colored.red("Not paused", vmrun))
+                    else:
+                        print(colored.yellow("Paused", vmrun))
                 else:
-                    print(colored.yellow("Paused", vmrun))
+                    print(colored.red("Not yet implemented on this platform."))
             else:
                 print(colored.red("VM ({}) not created.".format(instance)))
 
@@ -628,15 +650,18 @@ class Mech(MechCommand):
             inst = MechInstance(instance)
 
             if inst.created:
-                vmrun = VMrun(inst.vmx)
-                state = vmrun.check_tools_state(quiet=True)
-                if state == "running":
-                    print("VM must be stopped before doing upgrade.")
-                else:
-                    if vmrun.upgradevm(quiet=False) is None:
-                        print(colored.red("Not upgraded", vmrun))
+                if inst.provider == 'vmware':
+                    vmrun = VMrun(inst.vmx)
+                    state = vmrun.check_tools_state(quiet=True)
+                    if state == "running":
+                        print("VM must be stopped before doing upgrade.")
                     else:
-                        print(colored.yellow("Upgraded", vmrun))
+                        if vmrun.upgradevm(quiet=False) is None:
+                            print(colored.red("Not upgraded", vmrun))
+                        else:
+                            print(colored.yellow("Upgraded", vmrun))
+                else:
+                    print(colored.red("Not yet implemented on this platform."))
             else:
                 print(colored.red("VM ({}) not created.".format(instance)))
 
@@ -670,46 +695,49 @@ class Mech(MechCommand):
             # if we have started this instance before, try to unpause
             if inst.created:
 
-                vmrun = VMrun(inst.vmx)
-
-                if vmrun.unpause(quiet=True) is not None:
-                    print(colored.blue("Getting IP address..."))
-                    ip_address = inst.get_ip(wait=True)
-                    if not disable_shared_folders:
-                        utils.share_folders(vmrun, inst)
-                    else:
-                        print(colored.blue("Disabling shared folders..."))
-                        vmrun.disable_shared_folders(quiet=False)
-                    if ip_address:
-                        print(colored.green("VM resumed on {}".format(ip_address)))
-                    else:
-                        print(colored.green("VM resumed on an unknown IP address"))
-
-                else:
-                    # Otherwise try starting
+                if inst.provider == 'vmware':
                     vmrun = VMrun(inst.vmx)
-                    started = vmrun.start()
-                    if started is None:
-                        print(colored.red("VM not started"))
-                    else:
+
+                    if vmrun.unpause(quiet=True) is not None:
                         print(colored.blue("Getting IP address..."))
                         ip_address = inst.get_ip(wait=True)
                         if not disable_shared_folders:
                             utils.share_folders(vmrun, inst)
-                        if ip_address:
-                            if started:
-                                print(colored.green("VM ({}) started on "
-                                                    "{}".format(instance, ip_address)))
-                            else:
-                                print(colored.yellow("VM ({}) already was started "
-                                                     "on {}".format(instance, ip_address)))
                         else:
-                            if started:
-                                print(colored.green("VM ({}) started on an unknown "
-                                                    "IP address".format(instance)))
+                            print(colored.blue("Disabling shared folders..."))
+                            vmrun.disable_shared_folders(quiet=False)
+                        if ip_address:
+                            print(colored.green("VM resumed on {}".format(ip_address)))
+                        else:
+                            print(colored.green("VM resumed on an unknown IP address"))
+
+                    else:
+                        # Otherwise try starting
+                        vmrun = VMrun(inst.vmx)
+                        started = vmrun.start()
+                        if started is None:
+                            print(colored.red("VM not started"))
+                        else:
+                            print(colored.blue("Getting IP address..."))
+                            ip_address = inst.get_ip(wait=True)
+                            if not disable_shared_folders:
+                                utils.share_folders(vmrun, inst)
+                            if ip_address:
+                                if started:
+                                    print(colored.green("VM ({}) started on "
+                                                        "{}".format(instance, ip_address)))
+                                else:
+                                    print(colored.yellow("VM ({}) already was started "
+                                                         "on {}".format(instance, ip_address)))
                             else:
-                                print(colored.yellow("VM ({}) already was started on an "
-                                                     "unknown IP address".format(instance)))
+                                if started:
+                                    print(colored.green("VM ({}) started on an unknown "
+                                                        "IP address".format(instance)))
+                                else:
+                                    print(colored.yellow("VM ({}) already was started on an "
+                                                         "unknown IP address".format(instance)))
+                else:
+                    print(colored.red("Not yet implemented on this platform."))
             else:
                 print(colored.red("VM not created"))
 
@@ -735,11 +763,14 @@ class Mech(MechCommand):
             inst = MechInstance(instance)
 
             if inst.created:
-                vmrun = VMrun(inst.vmx)
-                if vmrun.suspend() is None:
-                    print(colored.red("Not suspended", vmrun))
+                if inst.provider == 'vmware':
+                    vmrun = VMrun(inst.vmx)
+                    if vmrun.suspend() is None:
+                        print(colored.red("Not suspended", vmrun))
+                    else:
+                        print(colored.green("Suspended", vmrun))
                 else:
-                    print(colored.green("Suspended", vmrun))
+                    print(colored.red("Not yet implemented on this platform."))
             else:
                 print("VM has not been created.")
 
@@ -898,7 +929,10 @@ class Mech(MechCommand):
             inst = MechInstance(instance)
 
             if inst.created:
-                utils.provision(inst, show)
+                if inst.provider == 'vmware':
+                    utils.provision(inst, show)
+                else:
+                    print(colored.red("Not yet implemented on this platform."))
             else:
                 print("VM not created.")
 
@@ -924,29 +958,34 @@ class Mech(MechCommand):
             inst = MechInstance(instance)
 
             if inst.created:
-                vmrun = VMrun(inst.vmx)
 
-                print(colored.blue("Reloading machine..."))
-                started = vmrun.reset()
-                if started is None:
-                    print(colored.red("VM not restarted"))
-                else:
-                    print(colored.blue("Getting IP address..."))
-                    ip_address = inst.get_ip(wait=True)
-                    if ip_address:
-                        if started:
-                            print(colored.green("VM ({}) started "
-                                                "on {}".format(instance, ip_address)))
-                        else:
-                            print(colored.yellow("VM ({}) already was started on "
-                                                 "{}".format(instance, ip_address)))
+                if inst.provider == 'vmware':
+
+                    vmrun = VMrun(inst.vmx)
+
+                    print(colored.blue("Reloading machine..."))
+                    started = vmrun.reset()
+                    if started is None:
+                        print(colored.red("VM not restarted"))
                     else:
-                        if started:
-                            print(colored.green("VM ({}) started on an unknown IP "
-                                                "address".format(instance)))
+                        print(colored.blue("Getting IP address..."))
+                        ip_address = inst.get_ip(wait=True)
+                        if ip_address:
+                            if started:
+                                print(colored.green("VM ({}) started "
+                                                    "on {}".format(instance, ip_address)))
+                            else:
+                                print(colored.yellow("VM ({}) already was started on "
+                                                     "{}".format(instance, ip_address)))
                         else:
-                            print(colored.yellow("VM ({}) already was started "
-                                                 "on an unknown IP address".format(instance)))
+                            if started:
+                                print(colored.green("VM ({}) started on an unknown IP "
+                                                    "address".format(instance)))
+                            else:
+                                print(colored.yellow("VM ({}) already was started "
+                                                     "on an unknown IP address".format(instance)))
+                else:
+                    print(colored.red("Not yet implemented on this platform."))
             else:
                 print("VM not created.")
 
@@ -976,16 +1015,19 @@ class Mech(MechCommand):
         for instance in instances:
             inst = MechInstance(instance)
 
-            print('Instance ({}):'. format(instance))
-            nat_found = False
-            vmrun = VMrun(inst.vmx)
-            for line in vmrun.list_host_networks().split('\n'):
-                network = line.split()
-                if len(network) > 2 and network[2] == 'nat':
-                    print(vmrun.list_port_forwardings(network[1]))
-                    nat_found = True
-            if not nat_found:
-                print(colored.red("Cannot find a nat network"), file=sys.stderr)
+            if inst.provider == 'vmware':
+                print('Instance ({}):'. format(instance))
+                nat_found = False
+                vmrun = VMrun(inst.vmx)
+                for line in vmrun.list_host_networks().split('\n'):
+                    network = line.split()
+                    if len(network) > 2 and network[2] == 'nat':
+                        print(vmrun.list_port_forwardings(network[1]))
+                        nat_found = True
+                if not nat_found:
+                    print(colored.red("Cannot find a nat network"), file=sys.stderr)
+            else:
+                print(colored.red("Not yet implemented on this platform."))
 
     def list(self, arguments):
         """
